Value Matching

This operator is simply the inverse of the native %in% operator, and will return true if the left-hand side is not in the right hand side. See the examples for more details.

x %nin% table

Arguments

x
vector or NULL; the values to be checked.

table
vector or NULL, the vectors to be checked against.

Value

A logical vector indicating if a match was not located for each element of x.

Examples

best_cats <- c("Tucker", "Oliver", "Mr. Loon")

# Returns TRUE, obviously
"Tucker" %in% best_cats

#> [1] TRUE

# Returns FALSE (because Tucker is the best cat)
"Tucker" %nin% best_cats

#> [1] FALSE

# Create a vector of vowels
vowels <- c("a", "e", "i", "o", "u")

# Create a sentence
sentence <- "Sushi is great!"
# Split sentence into individual letters
sentence_letters <- strsplit(sentence, "")[[1]]

# Subset the sentence for all vowels
sentence_letters[sentence_letters %in% vowels]

#> [1] "u" "i" "i" "e" "a"

# Subset the sentence for all consonants
sentence_letters[sentence_letters %nin% vowels]

#>  [1] "S" "s" "h" " " "s" " " "g" "r" "t" "!"

# Vector of numbers 1 through 10
all_nums <- seq(0,10)
# Vector of a few numbers between 1 through 10
sub_nums <- c(1,3,6)

# Which numbers between 1 and 10 are in the subset
all_nums[all_nums %in% sub_nums]

#> [1] 1 3 6

# Which numbers between 1 and 10 are not in the subset
all_nums[all_nums %nin% sub_nums]

#> [1]  0  2  4  5  7  8  9 10